requirements.txt
aiomysql==0.0.20
aliyun-python-sdk-core==2.13.10
aliyun-python-sdk-core-v3==2.13.10
aliyun-python-sdk-kms==2.8.0
APScheduler==3.6.1
asn1crypto==1.0.1
asyncio==3.4.3
backports.lzma==0.0.14
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
configobj==5.0.6
configparser==4.0.2
crcmod==1.7
cryptography==2.7
decorator==4.4.1
dill==0.3.1.1
dominate==2.4.0
etelemetry==0.1.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Moment==0.9.0
Flask-MySQLdb==0.2.0
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
httplib2==0.16.0
idna==2.8
imgkit==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lxml==4.4.2
MarkupSafe==1.1.1
multiprocess==0.70.9
mysqlclient==1.4.4
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.0
nipype==1.4.0
numpy==1.17.2
odps==3.5.1
oss2==2.8.0
packaging==20.0
pandas==0.25.1
Pillow==7.0.0
plotly==4.4.1
prettytable==0.7.2
prov==1.5.3
psutil==5.6.5
pycparser==2.19
pycryptodome==3.9.0
pydot==1.4.1
pydotplus==2.0.2
pyecharts==1.6.2
PyMySQL==0.9.2
pyodps==0.8.3
pyparsing==2.4.6
python-dateutil==2.8.0
pytz==2019.3
pyxnat==1.2.1.0.post3
rdflib==4.2.2
redis==3.3.8
reportlab==3.5.34
requests==2.22.0
retrying==1.3.3
scipy==1.4.1
simplejson==3.17.0
six==1.12.0
snapshot-phantomjs==0.0.3
soupsieve==1.9.5
SQLAlchemy==1.3.10
traits==5.2.0
tzlocal==2.0.0
urllib3==1.25.6
uWSGI==2.0.18
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
XlsxWriter==1.2.6
